+++ app/_layout.tsx
@@ -32,7 +32,7 @@ export default function RootLayout() {
         <Stack>
           <Stack.Screen name="(auth)" options={{ headerShown: false }} />
           <Stack.Screen name="index" options={{ headerShown: false }} />
-          <Stack.Screen name="profile" options={{ headerShown: false }} />
+          <Stack.Screen name="my-profile" options={{ headerShown: false }} />
           <Stack.Screen name="select-photo" options={{ headerShown: false }} />
         </Stack>
       </QueryProvider>
