+++ app/my-profile/main.tsx
@@ -3,9 +3,9 @@ import { View, Text } from 'react-native';
 import { SafeAreaView } from "react-native-safe-area-context";
 import { ScrollView } from "react-native-gesture-handler";
 
-import ProfileImageBox from '@/components/profile/common/profile-image-box';
-import UserActivitySummary from '@/components/profile/common/user-activity-summary';
-import ControlButtons from '@/components/profile/my-profile/control-buttons';
+import ProfileImageBox from '@/components/my-profile/common/profile-image-box';
+import UserActivitySummary from '@/components/my-profile/common/user-activity-summary';
+import ControlButtons from '@/components/my-profile/main/control-buttons';
 
 // TODO: 데스트 데이터
 const userInfo = {
