+++ components/my-profile/create/name-input.tsx
@@ -1,7 +1,7 @@
 import React, { useImperativeHandle, useState, forwardRef } from 'react';
 import { router } from 'expo-router';
 import { useProfileStore } from '@/store/profile';
-import { Input } from '@/components/profile/create/input';
+import { Input } from '@/components/my-profile/create/input';
 
 const MIN_LENGTH = 2;
